setup: update route for project
