+++ src/App.js
@@ -1,14 +1,30 @@
+import { useState } from "react";
+import { QueryClient, QueryClientProvider } from "react-query";
+import { BrowserRouter } from "react-router-dom";
 import "./App.css";
-import Header from "./components/header/Header";
-import Body from "./components/body/Body";
+import { ReactQueryDevtools } from "react-query/devtools";
+import Root from "./root";
 
 function App() {
-  return (
-    <div className="App">
-      {/* <Header /> */}
-      <Body />
-    </div>
-  );
+    const [queryClient] = useState(
+        () =>
+            new QueryClient({
+                defaultOptions: {
+                    queries: {
+                        refetchOnWindowFocus: false,
+                    },
+                },
+            })
+    );
+
+    return (
+        <QueryClientProvider client={queryClient}>
+            <BrowserRouter>
+                <Root />
+            </BrowserRouter>
+            <ReactQueryDevtools position={"bottom-right"} />
+        </QueryClientProvider>
+    );
 }
 
 export default App;
