+++ src/pages/AdvancedSearch/index.jsx
@@ -1,8 +1,13 @@
 import React from 'react'
+import Header from '../../components/header/Header'
+import Body from '../../components/body/Body'
 
 const AdvancedSearch = () => {
   return (
-    <div>AdvancedSearch</div>
+    <div className="App">
+      {/* <Header /> */}
+      <Body />
+    </div>
   )
 }
